removing warning act 4 not very proud of changes

+++ src/pdf-merger/FilterPredictor.cpp
@@ -59,14 +59,14 @@ FilterPredictor::~FilterPredictor()
 std::string FilterPredictor::getDictionaryContentStr(std::string & in, size_t &pos )
 {
    size_t beg = in.find(DICT_START_TOKEN,pos);
-   if( beg == -1 )
+   if((int) beg == -1 )
    {
       return "";
    }
 
    beg += DICT_START_TOKEN.size();
    size_t end = in.find(DICT_END_TOKEN,beg);
-   if( end == -1 )
+   if((int) end == -1 )
    {
       return "";
    }
@@ -92,7 +92,7 @@ void FilterPredictor::obtainDecodeParams(Object *objectWithStream, std::string &
    for(; it != params.end();it++)
    {
       size_t pos = dictStr.find((*it).first);
-      if( pos != -1 )
+      if((int) pos != -1 )
       {
          pos += (*it).first.size();
 
@@ -123,7 +123,7 @@ void FilterPredictor::initialize(Object *objectWithStream)
       objectWithStream->getHeader(content);
       // we need to parse the header of file to obtain the decoder parameter      
       size_t position = content.find(DECODE_PARAM_TOKEN);
-      if(  position != -1)
+      if((int) position != -1)
       {
          position += DECODE_PARAM_TOKEN.size();
          std::string dictStr = getDictionaryContentStr(content,position);
